remove draft style, extract droppable
